Revert "Merge branch 'ThrowThings' into tmp"
This reverts commit 4579404a7550bbd62a296dcf4771eb050219552e, reversing
changes made to 9e937fd16af19ac3e50db54eb117559062cbc6a0.

+++ GamePrograming/Application/Source/Game/Houki.cpp
@@ -1,8 +1,8 @@
 ﻿/******************************************************
-* Houki.cpp	?
-* ?????F?C?T?T?g??
-* ?쐬???F2024/12/05
-* ?ŏI?X?V???F2024/12/05
+* Houki.cpp	箒
+* 制作者：イササトル
+* 作成日：2024/12/05
+* 最終更新日：2024/12/05
 *******************************************************/
 #include "framework.h"
 #include "DirectX/DirectX.h"
@@ -11,41 +11,41 @@
 #include "Houki.h"
 
 /****************************************************
-* ⴏ?????
+* 箒初期化
 *****************************************************/
 Houki::Houki(float x, float y, float r) : ThrowObject(x, y, r) {
-	//?e?N?X?`???ݒ?
+	//テクスチャ設定
 	m_uv.x = 0.3f;
 	m_uv.y = 0.05f;
 	m_texSize.x = 0.5f;
 	m_texSize.y = 0.9f;
 
-	//?T?C?Y
+	//サイズ
 	float aspect = m_texSize.x / m_texSize.y;
 	float height = 120.0f;
 	m_size = XMFLOAT2(height * aspect, height);
 
-	//?|?W?V?????ϊ?
+	//ポジション変換
 	b2Vec2 b2pos = Physics::ConvertDXtoB2Float2(m_pos);
-	//?{?f?B?쐬
+	//ボディ作成
 	Physics::CreateBody(&m_body, b2pos.x, b2pos.y, r, true, this);
 
-	//?T?C?Y?ϊ?
+	//サイズ変換
 	b2Vec2 b2size = Physics::ConvertDXtoB2Float2(m_size);
-	//?????蔻????
+	//当たり判定作成
 	Physics::CreateFixture(&m_body, b2size.x, b2size.y, 1.0f);
 
-	//?e?N?X?`??
+	//テクスチャ
 	m_tex.Load(L"Data/Texture/Houki.png");
 
-	//?d??
+	//重量
 	m_weight = WEIGHT_NORMAL;
 
 
 }
 
 /****************************************************
-* ⴏI??
+* 箒終了
 *****************************************************/
 Houki::~Houki() {
 
@@ -65,8 +65,8 @@ void Houki::Update(){
 	}
 	if (GetIsThrow()&&downFirst)
 	{
-		int a = m_pos.x - m_throwPos.x;
-		int b = m_pos.y - m_throwPos.y;
+		int a = static_cast<int>(m_pos.x - m_throwPos.x);
+		int b = static_cast<int>(m_pos.y - m_throwPos.y);
 		int c = a * a + b * b;
 
 		int downDistance = 700;
@@ -80,3 +80,6 @@ void Houki::Update(){
 	}
 
 }
+
+
+
